requirements.txt
torch
torchvision
onnx
onnxscript
onnxruntime
pillow
numpy
matplotlib
